clean up postgres/redshift split
